requirements.txt
# FastAPI server
fastapi
uvicorn
python-dotenv

# LLM APIs
google-generativeai

# Document Processing
requests
beautifulsoup4
html2text
python-docx
nltk
PyMuPDF>=1.21.0
lxml

# Vector Search & Keyword Search
faiss-cpu
numpy
rank_bm25

# Caching
cachetools
